refactor(debug): centralize ai perception audit output
