+++ Source/Portfolio/Controller/CAIController.cpp
@@ -154,8 +154,8 @@ void ACAIController::UninitializeControllerRuntime()
 	ClearBlackboardValues();
 	UnbindPerceptionEvents();
 
-	PrintPerceptionCandidateAuditSummary();
-	PrintBlackboardEngageLatencyAuditSummary();
+	FAIPerceptionDebug::PrintPerceptionCandidateAuditSummary(ControlledPawn_Cached, PerceptionCandidateAuditState);
+	FAIPerceptionDebug::PrintBlackboardEngageLatencyAuditSummary(ControlledPawn_Cached, BlackboardEngageLatencyAuditState);
 
 	ClearBlackboardEngageLatencyAudit();
 	ClearPerceptionCandidateAudit();
@@ -682,77 +682,3 @@ void ACAIController::RecordEngageAssignmentResolvedForAudit(AActor* InTargetActo
 	BlackboardEngageLatencyAuditState.FirstEngageAssignmentTargetActor = InTargetActor;
 }
 
-// Profiling Debug
-
-void ACAIController::PrintPerceptionCandidateAuditSummary() const
-{
-	if (!PerceptionCandidateAuditState.bEnabled) return;
-
-	const bool bHasRawPerception = PerceptionCandidateAuditState.FirstRawPerceptionTime >= 0.f;
-	const bool bHasValidTarget = PerceptionCandidateAuditState.FirstValidTargetTime >= 0.f;
-
-	const float firstRawLatency = bHasRawPerception
-		? PerceptionCandidateAuditState.FirstRawPerceptionTime - PerceptionCandidateAuditState.RuntimeStartTime
-		: -1.f;
-
-	const float firstValidLatency = bHasValidTarget
-		? PerceptionCandidateAuditState.FirstValidTargetTime - PerceptionCandidateAuditState.RuntimeStartTime
-		: -1.f;
-
-	FLog::Log(FString::Printf(
-		TEXT("[PerceptionCandidateAudit] Owner=%s | RawEvents=%d | RawActors=%d | ValidProviders=%d | InvalidProviders=%d | MaxTargetDataMap=%d | FirstRawLatency=%.3f | FirstValidLatency=%.3f | StartFrame=%llu | FirstRawFrame=%llu | FirstValidFrame=%llu"),
-		*GetNameSafe(ControlledPawn_Cached),
-		PerceptionCandidateAuditState.RawPerceptionEventCount,
-		PerceptionCandidateAuditState.RawPerceptionActors.Num(),
-		PerceptionCandidateAuditState.ValidTargetProviderActors.Num(),
-		PerceptionCandidateAuditState.InvalidTargetProviderActors.Num(),
-		PerceptionCandidateAuditState.MaxTargetDataMapSize,
-		firstRawLatency,
-		firstValidLatency,
-		PerceptionCandidateAuditState.RuntimeStartFrame,
-		PerceptionCandidateAuditState.FirstRawPerceptionFrame,
-		PerceptionCandidateAuditState.FirstValidTargetFrame));
-}
-
-void ACAIController::PrintBlackboardEngageLatencyAuditSummary() const
-{
-	if (!BlackboardEngageLatencyAuditState.bEnabled) return;
-
-	const bool bHasPerceptionContext = BlackboardEngageLatencyAuditState.FirstPerceptionContextTime >= 0.f;
-	const bool bHasBlackboardTarget = BlackboardEngageLatencyAuditState.FirstBlackboardTargetTime >= 0.f;
-	const bool bHasEngageRequest = BlackboardEngageLatencyAuditState.FirstEngageRequestTime >= 0.f;
-	const bool bHasEngageAssignment = BlackboardEngageLatencyAuditState.FirstEngageAssignmentTime >= 0.f;
-
-	const float perceptionContextLatency = bHasPerceptionContext
-		? BlackboardEngageLatencyAuditState.FirstPerceptionContextTime - BlackboardEngageLatencyAuditState.RuntimeStartTime
-		: -1.f;
-
-	const float blackboardTargetLatency = bHasBlackboardTarget
-		? BlackboardEngageLatencyAuditState.FirstBlackboardTargetTime - BlackboardEngageLatencyAuditState.RuntimeStartTime
-		: -1.f;
-
-	const float engageRequestLatency = bHasEngageRequest
-		? BlackboardEngageLatencyAuditState.FirstEngageRequestTime - BlackboardEngageLatencyAuditState.RuntimeStartTime
-		: -1.f;
-
-	const float engageAssignmentLatency = bHasEngageAssignment
-		? BlackboardEngageLatencyAuditState.FirstEngageAssignmentTime - BlackboardEngageLatencyAuditState.RuntimeStartTime
-		: -1.f;
-
-	FLog::Log(FString::Printf(
-		TEXT("[BlackboardEngageLatencyAudit] Owner=%s | PerceptionContextLatency=%.3f | BlackboardTargetLatency=%.3f | EngageRequestLatency=%.3f | EngageAssignmentLatency=%.3f | StartFrame=%llu | PerceptionContextFrame=%llu | BlackboardTargetFrame=%llu | EngageRequestFrame=%llu | EngageAssignmentFrame=%llu | PerceptionTarget=%s | BlackboardTarget=%s | EngageRequestTarget=%s | EngageAssignmentTarget=%s"),
-		*GetNameSafe(ControlledPawn_Cached),
-		perceptionContextLatency,
-		blackboardTargetLatency,
-		engageRequestLatency,
-		engageAssignmentLatency,
-		BlackboardEngageLatencyAuditState.RuntimeStartFrame,
-		BlackboardEngageLatencyAuditState.FirstPerceptionContextFrame,
-		BlackboardEngageLatencyAuditState.FirstBlackboardTargetFrame,
-		BlackboardEngageLatencyAuditState.FirstEngageRequestFrame,
-		BlackboardEngageLatencyAuditState.FirstEngageAssignmentFrame,
-		*GetNameSafe(BlackboardEngageLatencyAuditState.FirstPerceptionTargetActor.Get()),
-		*GetNameSafe(BlackboardEngageLatencyAuditState.FirstBlackboardTargetActor.Get()),
-		*GetNameSafe(BlackboardEngageLatencyAuditState.FirstEngageRequestTargetActor.Get()),
-		*GetNameSafe(BlackboardEngageLatencyAuditState.FirstEngageAssignmentTargetActor.Get())));
-}
